internal/reflectlite: do not use the reflect package
This avoids a dependency on the reflect package.

I've tried to deduplicate as much as reasonably possible, while keeping
the code clean. Especially some tricky methods for the reflect.Type type
have a shared implementation (using //go:linkname). In other cases, I've
simply used the same type using a type alias, in the other direction
(the reflect package now imports the internal/reflectlite package
instead of the other way around).
